[ancillary.combine_polygons] moved from search.combine_polygons
